+++ frontend/src/stores/google-maps/googleMapStore.ts
