+++ frontend/src/stores/google-maps/stationSummaryWindowStore.ts
