--- frontend/src/stores/navItemsOpenStore.ts
@@ -1,5 +1,0 @@
-import { store } from '@utils/external-state';
-
-export const stationSearchWindowOpenStore = store(false);
-
-export const serverStationFiltersOpenStore = store(false);
